Thread in "on demand" edit distance calculation.  Had to spend some time figuring out what the code I wrote several months ago acatually did.
Sed.pm now supports two modes of operation.  Calling get_sed($id1, $id2), will force it into
one of the following modes, depending on the revelant conditions:

  1)The distance matrix hash is already loaded, and if not, whether the file
    exists on disk.  If either of these conditions are true, the distance matrix is
    loaded from disk, if necessary, and these pre-computed valued are used to serve the query.
    The SeD object keeps the distance matrix hash in memory until the SeD object is freedd.

  2)The distance matrix hash is not loaded and the corresponding file containing distances
    does exist on disk.  In this case, the SeD object loads the input_vector file, which contains
    the string representation of each category or cluster into memory.
    The object indexes into the input_vector to retrieve the appropriate strings, and computes
    the edit distance "on demand."  The input_vector is kept in memory until the object is freed,
    or until "calculate_all_edit_distances()" is called.

On demand edit distance calculation is now the default mode of operation.
Sometimes, for experimental purposes, it is still useful to calculate distances
between all of categories.  In such cases, the user can override the default
behaviour by specifying --calc_all_edit_dists on the command line.

The --bypass_sed option is still supported and will work for both modes.

The rest of the changes in this commit were necessary to thread the new mode of
SeD operation through the code.  Most notably CreateClusteringInput no longer
calculates SeDs internally; Rather, a SeD object is created by the top-level
Spectroscope.pl script and passed to the modules that need to compute SeDs
(ParseClusteringResults, IdentifyMutations, etc).

+++ source/lib/PassThrough/PassThrough.pm
@@ -1,6 +1,6 @@
 #! /usr/bin/perl -w
 
-# $cmuPDL: PassThrough.pm,v 1.3 2009/07/27 20:08:21 rajas Exp $
+# $cmuPDL: PassThrough.pm,v 1.4 2009/08/06 17:33:20 rajas Exp $
 ##
 # This perl modules implements a "ClusteringPassThrough"
 # that is, it generates as many clusters as input
@@ -49,8 +49,8 @@ my $_remove_existing_files = sub {
 #### API functions ###############################
 
 sub new {
-    assert(scalar(@_) == 3);
-    my ($proto, $input_dir, $output_dir) = @_;
+    assert(scalar(@_) == 4);
+    my ($proto, $input_dir, $output_dir, $sed) = @_;
 
     my $class = ref($proto) || $proto;
 
@@ -58,7 +58,7 @@ sub new {
 
     # Input files
     $self->{INPUT_VECTOR_FILE} = "$input_dir/input_vector.dat";
-    $self->{INPUT_DISTANCE_MATRIX_FILE} = "$input_dir/input_vector_distance_matrix.dat";
+    $self->{INPUT_DISTANCE_MATRIX_FILE} = $sed->get_distance_matrix_file();
 
     # Output files
     $self->{CLUSTER_DISTANCE_MATRIX_FILE} = "$input_dir/clusters_distance_matrix.dat";
@@ -107,9 +107,11 @@ sub cluster {
         $offset++;
     }
 
-    system("cp $self->{INPUT_DISTANCE_MATRIX_FILE} " .
-           "$self->{CLUSTER_DISTANCE_MATRIX_FILE}") == 0
-           or die "Could not create $self->{CLUSTER_DISTANCE_MATRIX_FILE}";
+    if( -e $self->{INPUT_DISTANCE_MATRIX_FILE}) { 
+        system("cp $self->{INPUT_DISTANCE_MATRIX_FILE} " .
+               "$self->{CLUSTER_DISTANCE_MATRIX_FILE}") == 0
+               or die "Could not create $self->{CLUSTER_DISTANCE_MATRIX_FILE}";
+    }
 
     close($input_vector_fh);
     close($output_fh);
